wmcs.backup_glance_images: disable the backups on 1003 and 1004
Cloudcontrol1003 and 1004 have spinning disks, and doing the backups,
even if the data does not change, rewrites all the metadata to the
local disk:

https://github.com/wamdam/backy2/issues/3

And it turns out that those two servers have spinning disks and this
overloads them quite significantly (see the related task).

This disables the backups on those two servers, and adds a variable to
fine control which of the control nodes actually do backups.

Bug: T270478
Change-Id: I2f3d504e9f6e0869c685fdb049d8d00bfd1cd523
Signed-off-by: David Caro <dcaro@wikimedia.org>