wikireplicas: enable standby behavior on multiinstance proxies
This is for the physical server (production realm) proxies. The intent
is to generate the hash for the other "replica_type" with an additional
hash key of "standby" to generate appropriate config from.

Standby servers would still be removed by the inclusion of the server in
the overrides hiera with a "depooled" key.

Since the config doesn't interfere with the legacy systems on the proxy
servers, we should be pretty safe to play with this and the overrides
unless puppet just fails.

Bug: T271476
Change-Id: Iddd46a82a7505f95e9caee89340e21d2e0345f41