wmcs.sge.prometheus Retry getting the job count
We are getting some mails for failed crons, it seems that when trying
to read the job count the file might be empty, and the script just
crashes trying to cast to integer.

The guess is that sge is in-between updating the file with the new
count.

This adds a couple retries to give time for sge to update the file
correctly and should make the script more resilient giving the correct
result more often. Might have to tweak the amount of retries and/or
the sleep time.

Note that the SGE master node runs deprecated python 3.5 that does not
support f-strings.

== Test plan

Uploaded and ran on the sge master node:
```
dcaro@vulcanus$ \
    ssh tools-sgegrid-master.tools.eqiad1.wikimedia.cloud \
        ./prometheus-sge-stats.py | arc paste --
 DONE  Created a new paste.
  *   P13710 Command-Line Input
 ://    https://phabricator.wikimedia.org/P13710
```

Bug: T271686
Change-Id: I3ae4f452c65b0ecbdb89fa52f0dfc73150b74c42
Signed-off-by: David Caro <dcaro@wikimedia.org>