ores: Switch from oresrdb.svc to host names
Why:
The idea never really panned out after all. It was there to overcome the
limitation of ORES being unable to fully be behind nutcracker and thus
fully HA. However the fallback process requires anyway a manual slow
restart of ORES in the case of a scheduled migration and a manual faster
restart in the case of an emergency, both after a gerrit change has been
merged. This led to a situation where, operationally speaking, it hasn't
really been better than just switching the host used and restarting
everything (slowly or not).

What:
Switch to hardcoding the hostnames. While it seems like a step back, it
isn't as the operational burden remains the same and at least 1 moving
part is removed

Bug: T270071
Change-Id: Id479cdc19bee1cc3d3f2f3e0c3c1fc6c54f2f614