Reapply Ib32827c9433f362600643666f0156fed567fb3bc
Current lvm provider code does not delete a volume group if that volume
group has logical volumes assigned to it and fails badly. While this
might appear as a safety precaution, it's breaking puppet runs while not
really helping the user since they probably know what they are doing if
they want to remove a volume group. Always force removal instead

Note that this is one of the very rare exceptions where we have diverged
from an upstream module. The main reason for that is that we intended to
drop the module once we moved away from Docker LVM, which is still bound
to happen, albeit no ETA yet.

Bug: T271099
Change-Id: Ia4b818ff6e39e42f5ba2614af210ed2cdfb22f99