wikireplicas: set up LVS for multiinstance wikireplicas
The replicas do not need any load balancing at the LVS layer. They are
load balanced by haproxy (eventually 2 sets of them). The goal here is
compliance with production network expectations stemming from T267376
and on-going discussions with network folks about how Cloud VPS VMs
access services that are served from production networks.

The aim is to force the VM services to access the wikireplicas in an
expected fashion. If this should be simply through a public IP, we can
change to that, but this is my understanding of the requirement at this
time. No IPs have been provisioned yet for this, pending more review.

Bug: T271476
Bug: T272720
Change-Id: If49d66b64c1b8038bcdfda72843998159edf4d83