dnsrecusor: update variable name for installing version from component
The `enable_pdns43' variable decides the version of pdns-recursor that
is installed by the dnsrecursor module. We should update its name so
that it better reflects its actual purpose: if this is set to true (for
Wikidough), the version from component is installed. If set to false
(for currently everything other than Wikidough), the version in Debian
stable is installed.

Update the Wikidough profile to reflect this change.

This commit is to prepare updating pdns-recursor component from 4.3.3 to
4.4.2, the latest version in testing.

Change-Id: Ib899d90164c783ef1e7551324db96bbb66d6642f