nfs: change the default throttles for primary cluster read and egress
First, let me say that this only gets applied ever to NFS clients and now
NFS should all be capable of 10Gb Ethernet for reads. Writes are still
going to be affected by the 1Gb Ethernet crossover connection for DRBD.
That said, here's the math for people to consider it.

We currently have an average number of VMs per hypervisor of around
30. If you figure that every single VM is affected by these and flooding
their available limits (not something that would be normal at all), then
each would get 45 MB/s as their share of the 10Gbps connection. Here is
where things get confusing. MB as in mega-bytes per second. tc uses bytes.

So, the original egress of 30000kbps is actually 30 MB/s or 240Mbits/s. I am
upping that there so the back-and-forth isn't throttled anyway by that.

tc understands "mbps" as well as "kbps", so using that for nfs_read. That
should be understood to mean 80mbps in bits, which still feels slow to me
for a home directory, but it'll do! I would like to increase these values
on the Toolforge bastions horizon hiera to something higher so we are not making
those shared servers suck basically all the time.

FYI: unthrottled VMs (all non-NFS vms, basically) can achieve a max speed
of 9.38 Gbits/sec in real world testing right now. We hobble NFS vms a lot.

Also note that this change will only affect new VMs until tc-setup.sh is
run across VMs that have it.

Change-Id: I1ce9f54adce742a228087e2c3b62936a87d39d30