profile::analytics::refinery Create HDFS folders
A set of predefined HDFS folders with defined ownership are needed
in refinery to move data temporarily:
 - For druid indexation
 - For analytics oriented jobs (data-quality and archive)
The folders creation is activated using a hiera lookup flag. It is
set to false by default, and to true for analytics launcher role only.
Folders are: /wmf/tmp/analytics and /wmf/tmp/druid

Bug: T271560
Change-Id: I600b21ff0b9968d849f8ebcdfa78818ddf786480