maps: reimage maps1009 with buster.
Reimage maps1009 to buster to ensure that puppet completes and to
begin the process of reimaging it to a new independent
master. Reimaging this host will not cause it to resync its postgres
or to start cassandra without manual intervention so this operation
should not impact production maps traffic. maps1009 is currently not
pooled as a kartotherian host.

Bug: T238753
Change-Id: Iaecb9caf0d7d79f7e5456d89c5c60b2beb463b42