maintain-dbusers: skipping servers needs to skip closing too
We skip down servers in maintain mode, but now that we also close connections
correctly, we have to skip if that doesn't work as well.

In the near future it may be good if the service also emails to let us
know we *might* need to run a harvest_dbaccounts soon.

Bug: T272127
Change-Id: Ib912d88d6f85b93f092239001a3b3d94860e37e8