jenkins: allow path with encoded slashes
Matrix jobs can have parameters whose value contain a slash. The Jenkins
UI properly urlencode them in the URL however Apache 2.4 defaults to
decode them and reject such URL.  That left us with a 404 when trying to
reach the build pages for example on:

https://integration.wikimedia.org/ci/job/mediawiki-php-security/

Instruct Apache to allow the encoded slashes and prevent it from
decoding them so they get passed as-is to Jenkins.

Hosts: contint2001.wikimedia.org, releases2002.codfw.wmnet
Bug: T272159
Change-Id: I00b667b80d31154f625fa7a8e8834683a3eb72f6