maps: make maps1009 a new, independent buster master.
Once maps1009 has been reimaged as a buster host, make it an
independent master, segmented from the other hosts in eqiad's maps
cluster.

maps1009 will be a postgres master, the sole member of the cassandra
cluster and a redis master. Removing maps1009 from the general cluster
configuration for maps::master and maps::replica will cause the
firewall rules allowing maps1009 to be removed, thereby reducing the
risk of accidental overlap with the existing cluster.

maps1009 is currently not pooled in kartotherian and is not part of
the existing maps1.0 cassandra cluster.

Bug: T238753
Change-Id: I5589a4f2e9fe9f0e9148abf3ef01db4646ee1e70