Nova: add a simple vendordata REST service
Eventually we will replace nova's static vendordata file
with a call to this service. The advantage of using a service
over a static file is that the service can customize the vendordata
based on instance details (specifically, name and project).

The service is a paste/webob service the takes a vendordata jinja template.

Bug: T271273
Change-Id: I54d4b2910b19f2748e35160eb4c333c9467f367f