wikireplicas: Add new DNS names for multiinstance replicas
In order to open things up to early adopters, the new replicas need
a frontend. This should Just Work with the existing script if no delegations
or what-have-you are required. The IP addresses are labeled in neutron
and can be checked with "sudo wmcs-openstack --os-project-id clouddb-services port list"
on a cloudcontrol node. The ips are already attached to the proxies and
are live. The LVS they need to connect through are not yet ready, though.

You should be able to ping the relevant IPs at least, if not anything else.

Bug: T267376
Bug: T271476
Change-Id: I5a010da302db17f5899b046bb51d7b899903175c