Revert "Nova: add a simple vendordata REST service"
This reverts commit 6dc5f1dbb47630354f1a9b392fde7d791a7dc99e.

Reason for revert:

Since writing this, I have learned a few things:

- cloud-init totally ignores dynamic vendordata. That's
https://bugs.launchpad.net/cloud-init/+bug/1841104, which
I've submitted a patch for.

- recent versions of cloud-init have a built-in jinja
template engine. So rather than all the complexity of
this patch I'm building a new Stretch base image with
a modern cloud-init built in, at which point I should
be able to accomplish the VM-specific metadata things
I need with .erb and jinja and no external steps.

Change-Id: I343abcdd77543056e89cbdfe6726cec3f95e3706