varnish: block python-request UA bots for AQS
Some hours ago AQS got a ton of traffic to the pageview API from
a bot with python-requests/etc.., breaching our policy for UA hitting
APIs. The original block is extended to a more permanent one, that
will be curated by Analytics when needed (this is the proposal for
this patch).

Change-Id: I3ae906723919b4c26a767082c403186090700a5b