nrpe_check_opcache: use full path to php7adm to fix opcache monitor on buster
The opcache hit rate Icinga check did not work on buster appservers.

Running the script locally, also as the nagios user, worked on all
hosts.

But running it via NRPE showed different results on stretch vs buster.

All buster hosts would return "CRITICAL: opcache cache-hit ratio is below 99.85%"
to NRPE but "OK: opcache is healthy" when running the same command locally.

After debugging for a while how this is possible I eventually found this.

Even though php7adm is in the same location, on buster this fails
when not using the full path, which worked in stretch.

I tested it does not affect stretch hosts but fixes the issue on buster.
It does not even need an NRPE restart, just starts working immediately.

Bug: T270517
Change-Id: I2be08be4e54c00326839fbf92fef583154b7636f