profile::analytics::cluster::users: add analytics to the druid group
After the recent refactor of pemissions/ownership in HDFS (with
a new default umaks etc..) the Analytics team has been working on
fixing various jobs running on Hadoop that were previously assuming
a very lenient perm config. This change is needed to allow the
analytics user, running via Hive, to properly move and chgrp group
to directories created hourly/daily by batch jobs. We chose to avoid
solutions like the druid user in the analytics-privatedata-group to
limit the scope of what Druid can read on HDFS.

Change-Id: I2b3da650787cd28d19e6e91c058cbd279a4bbbc8