mediawiki: Port nrpe_check_opcache to Python
It's pretty likely the issue in T270517#6764034 would've been easier to
debug had it been written in a saner language where failing to shell out
would immediately raise an exception instead of being silently hidden.

So replace the combination of bash/awk/jq with a Python script which is
hopefully both easier to follow and debug.

Change-Id: I69a7d46f5aeaa3a0933325d9f55c4a91aad072a2