enable-puppet: allow fall back to enable puppet disabled by root
If puppet is disabled by root the disable message is not appended with
` - SUDO_USER`.  With the current enable-script its not possible to
re-enable puppet using sudo, as it assumes this postfix exists.

This issue most comes about when one disables puppet via cumin aka root
Then want to enable puppet on an individual server using sudo.

This CR adds a fall back to check for the enable message with and
without the postfix.

I opted to do the simple and call jq and grep twice.  I wanted to avoid
just checking for ''.startwith as this seemed error prone as did using
`grep -E` with a string built in bash. Happy to be convinced
otherwise or do something else

Bug: T272539
Change-Id: I36b843f968090bf398005df354d5fdfecbfe7d01