data-services: apply user variances to future creations
Since we now have 16 instances (soon to be 24) to create users across
in the wikireplicas service, we should set the changed max_connections
values so that they are created correctly during a rebuild rather than
going back and fixing manually.

This does not fix the problem of actually changing the value across all
instances retroactively now or in the future. It merely prevents rebuilds
from doing something incorrect. As is, the multi-instance cluster does
not have these variances applied yet, and that will need doing by hand.

Bug: T269399
Change-Id: Ib39c5c6ed274c6eacad19573ef4c354a5acf4636