reduce number of xml/sql dumps kept on dumpsdata hosts by one
We can reduce this for all but the small wikis with relatively low risk.
In a worst case scenario, we could always grab the files for a particular
wiki and earlier run from the labstore hosts.

This will keep dumpsdata1003 below the 95% mark for space used for a while
longer.

Change-Id: If87b364327b65baf07cbc2f70aba9829d7ab139a