Remove obsolete tmpreaper Puppet classes
This was once used on the app servers, but the tool has been showing it's
age and stopped using it (e.g. there were issues with tmpfiles and
systemd units using a private /tmp).

Toolforge still uses tmpreaper, but not the Puppet classes, it ships
a custom config file and installs the deb separately.

Bug: T272559
Change-Id: I19089c26160ad8e3f4cf65fa3948171218ff1135