wikireplicas: deploy a cloud-based query sampler for the replicas
In order to better understand the types of queries and traffic that hit
the wikireplicas services, create a service that can gather information
about those queries along with enough metadata to answer arbitrary
questions about it.

This should live on cinder to make the data somewhat portable and in
clouddb-services to make it easier to control access to the full,
unfiltered metadata.

Bug: T272723
Change-Id: I55b37479d873f8d68e3f301235c29bcfdb31226f