use the platform-engineering group to add people to deployment
...and remove their names from the explicit list.

When people move into the team we can add them to the group to give them
deployer rights and access to the dumps hosts, and anything else that
team members generally need.

When people move to another team we'll want to remove their names from
the platform-eng group and add them explicitly to the deployers list,
until more teams manage their access this way.

Note: removing gmodena and nikkin from the group because they do
not yet have deployment access, and would need to submit phab tasks
to get it.

Change-Id: I6dacd1a03f5e4371d5b6b03bca210de4d329a213