services: Create LVS services for linkrecommendation
Why:
linkrecommendation is almost ready to be exposed to mediawiki, setting
up LVS for it in this commit and follow ups. Alongside the internal LVS
service, setup and one on a different port that will be exposed publicly
via the edge caches, in the beginning primarily for PM and QA testers to
have access to it without having to jump through tunneling hooks
We don't expose publicly the internal one in order to follow the
separation of concerns principle. The second service is powered by a
different release on the kubernetes nodes, so, at least kubernetes
wise, capacity and users will be different.

What:
Create 2 LVS entries, albeit with the same IPs, just different ports.

Bug: T265603
Bug: T269581
Change-Id: Ib2a2478527be0b39b5e9327af71222e54b989644