archiva: disable nginx proxy buffering
Nginx proxy buffering is disabled on several of our system (for example
on the Docker registry).

Buffering is enabled by default with a default size of one memory page
(4k or 8k).  If the response does not fit in memory, Nginx uses
temporary files (default up to 1G) written in chunk of two memory pages
(8k or 16k).   I am guessing that lot of artifacts do not fit in memory
causing Nginx to write to disk on all requests?

Disabling the buffering means Nginx would instead start serving files
immediately and synchronously. Temporary files would no more be used.
The drawback is that for slow clients, the connections stay active for a
bit longer on the backend.  I am guessing it is not going to be a
problem.

That might speed up requests for https://archiva.wikimedia.org/

Some refs:
https://docs.nginx.com/nginx/admin-guide/web-server/reverse-proxy/#buffers
https://nginx.org/en/docs/http/ngx_http_proxy_module.html#proxy_buffer_size

Bug: T252767
Change-Id: I2162a1fb5cd07b9c837274f3cc7c1673139e8d13