Update minimum expected file size for lexeme JSON dumps
The script expects dumps to have a certain minimal size (otherwise
something went probably wrong), but that expected minimal size should
not be the same for the “all” (items and properties) and “lexeme”
(lexemes) dumps, since lexeme dumps are expected to be smaller. (The RDF
dumping already respects this: compare wikidatardf_functions.sh,
setDumpNameToMinSize().)

The minimum size of 100 megabytes (for all shards, so 12.5 megabytes per
shard) is mainly based on the error message quoted in T264883#6787346,
where a single shard was observed to have some 26 megabytes – I’m
assuming that all shards have roughly the same size. (The minimum size
for “all” dumps is left untouched, though it’s surely outdated by now –
I see no need to include the risk of updating it in this change.)

Also, use Bash arithmetic to calculate and compare the minimum size per
shard (already used e.g. for numberOfBatchesNeeded), and include the
expected minimum size in the error message for convenience.

Bug: T264883
Change-Id: Idc6df3372f117d52ab8165218841d37b42ac6624