dnsdist: update template to disable broken security polling
As discussed in T273322, the security polling for dnsdist is broken
because of an "invalid" version string that is generated during Debian
packaging; see the task above for more details.

Since this does not serve any purpose (the security status is not
checked) and shows up as an error in the logs, we disable this by
setting this option to an empty string.

This matches the configuration file in dnsdist's Debian package.

Bug: T273322
Change-Id: I3a05b121cd15283150a3ba9a07b1056203a8a9fb
Hosts: malmok.wikimedia.org