modules/network: Break out management networks by DC label
This change alters network/data/data.yaml so that the management
networks are labelled with their data center. The justification is
to allow some things that need to be generated for each data center's
management network (for example, parts of the automated DHCP
configuration) to be done in a more DRY manner. The network::constants
module is also modified so that it still provides the $mgmt_networks
variable unchanged that is used by other modules, but now provides the
additional $mgmt_networks_bydc variable which is a hash keyed by data
center.

Bug: T271583
Change-Id: Ia7c2e08b45b45304b87ba1cad3cd86e6ff2fadb0