refactor script for wikidata and commons rdf dumps
This is step one of making these jobs easier to test in deployment-prep.

We need to be able to specify alternative configuration files, which means
another command line argument. And that's too much to add into the already
complicated logic for processing values passed on the command line;
convert them all to proper arguments and flags.

Bug: T269377
Change-Id: Ie9c5c32cf18eeec17e54f7395c92960514cbecae