logging::mediawiki::udp2log: replace cron with timer
This replaces the mw-log-cleanup cron with a systemd timer
as part of a larger effort to replace all cron jobs with timers/services
which has various advantages.

We are still running the same script as before which runs the actual find
commands to clean up logs, so that did not change.

Now this can be handled with regular "systemctl status/start" commands
and will show up in "systemctl list-timers".

https://puppet-compiler.wmflabs.org/compiler1002/28222/

Bug: T273673
Change-Id: Iee204d1b6440cf3d59a914729c358626c42ec287