conftool: restore maps1009 to kartotherian pool
while maps1009 is for all purposes an individual host in its own
cluster, we still need to deploy to it. Readd it to the conftool group
so that we can still use scap with it. This change does not change any
firewall rules or readd maps1009 to any clusters. maps1009 will be
disabled with weight set to 0 until we're ready.

Change-Id: Id356abdc796717605d4d20149450db7af075ff1c