Add a new profile to install OpenLDAP client tools in production
We currently have a number of systems running in production which
install LDAP client tools via profile::ldap::client::labs. This
profile is targeted at instances in Toolforge/Cloud VPS and has
many settings and includes not relevant for production (such as support
for LDAP auth in PAM/NSS/SSSD).

This creates a separate profile which is trimmed down to the use of
OpenLDAP for production hosts. This also drops the obsolete
ldaplist tools/ldapsupportlib which isn't ported to Python 3
and relies on the central LDAP related central config settings in Hiera.

This profile will also serve as the base to narrow down the
acess to the read/write OpenLDAP instances (since most services
only need the readonly replicas behind LVS), since the use of the
RW servers need to be opted in (which is needed e.g. for the mwmaint
hosts where user changes happen, while services only reading from
LDAP can use the readonly servers).

Also introduce a new proxyagent setting in the central LDAP config
which allows to deduplicate various role-specific settings.

Change-Id: I19bda607e3f3a42d63b0c452a265aa738d26d6d8