cumin: make some hadoop aliases more granular
This change introduces more aliases that will be useful when upgrading
the Analytics cluster to Bigtop. The analytics-tool1001 runs an
old version of Hue package by Cloudera, and it will not be upgraded to
Bigtop (we'll likely/hopefully deprecate it), meanwhile it would be
great to upgrade one coordinator at the time (since some daemons may
start right after install).

Change-Id: I37596df825cdf79fa27d1264d45fb2e584833c82