md: Run checkarray on a random weekday of each month
We had issues with different ganeti hosts running checkarray at the same
time, leading to high latency of multiple instances of one etcd cluster
because they where unluckily distributed to those ganeti instances.

This changes the mdadm cronjob so that it distributes the array checks
over the first 28 days of each month to give us more variance in start
dates.
It sticks to a weekday (e.g. checks do not start on weekends) as well as
the start time of 5:57 AM (as that might still give a hint to people if
they see high I/O on hosts.

Bug: T273953
Change-Id: I9895b6d9f38a87128e23867ae119620eabd3dc1c