interface: update rps script to also set the number of queues via ethtool
We currently use the `interface::rps` define to set the number of queues
an interface has using ethtool via exec.  This CR updates the current
interface-rps script to also include the ability to set the number of
queues

I think this may allow use to clean up some of the functionality in the
interfaces module but also completely drop the numa fact however I haven't
checked if there are other use cases of this fact.

Hosts: cumin:R:interface::rps
Bug: T236208
Bug: T263578
Change-Id: I16f6fc410398ec851b863471ab3d80a5c04c4bba