ceph.osd: Allow setting the io scheduler of the osd disks
This had to include a small refactor to be able to specify what disks
are os disks (as that's not in the facts).

This will also set it to 'none' that is supported on all kernels, but
if we move to a newer kernel we will have to filter by kernel
specificly too (but let's handle that when it's needed).

Bug: T273791
Change-Id: I6febaf5abd7dfbc2ed2b37a82191a8f311195d22
Signed-off-by: David Caro <dcaro@wikimedia.org>