Keystone: a new (but still terrible) approach to making projectid==projectname
Previously this was done via a monkeypatch in wmfkeystonehooks. In older
keystone versions the hooks were loaded early enough that the patch was
in place when we needed it, but in Rocky and later the hooks are loaded
on demaind and so the monkeypatch didn't get applied before the project
was created.

This patch just overlays the affected file instead. It will be more trouble
to keep this current but it's also not as likely to fool us with
entry-point races.

Bug: T274165
Change-Id: If79052197eb47b44b2e3eef360ab3dd3652b6c9b