systemd: add data type for 'day of the week' in systemd timers/calendar
'day of the week' (weekday) could be expressed as an integer in
cron jobs (1 = Monday, etc).

But in systemd timers and more generally calendar events, it needs
to be a string, English name of the day, in full or abbreviated form.

Adding a type for that to be used with some timers that have a weekday
parameter to run "every Monday".

https://www.freedesktop.org/software/systemd/man/systemd.time.html

Change-Id: I151a38eee2e612ed416df22d8834b234427ea3a9