toolforge: add haproxy and nginx-ingress configuration for jobs.toolforge.org
This is a special case configuration to serve the jobs.toolforge.org API endpoint.
The patch includes a new nginx-ingress.yaml manifest to be loaded into kubernetes.

So, summary:
 * deploy a new haproxy configuration file, to load balance the new TCP port used by the jobs API.
 * deploy nginx-ingress configuration yaml for kubernetes. This file contains all the bits required
   to spin up another ingress controller in the toolforge k8s cluster. This new ingress controller
   is used to verify client-side TLS certificates and finally call toolforge-jobs-api web service
   that will be running in the k8s cluster.

This patch doesn't touch the dynamicproxy front proxy.

For now, to test the setup, I will use the FQDN 'jobs.svc.toolsbeta.eqiad1.wikimedia.cloud', which
is a CNAME to 'k8s.toolsbeta.eqiad1.wikimedia.cloud'. So requests to the FQDN will go directly to
haproxy.

Bug: T274139
Change-Id: Idd12b886df4e5d9b436807cdb62f7e3b4229d15b
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>