mariadb: Promote db1118 as the new candidate master for s1
On puppet this only means documentation and enabling STATEMENT
binlog on it in case if fails over. The main blocker is that
db1118 is upgraded to 10.4, which means if db1084 fails, we will have
a 10.4 source traffic for binlog, potentially breaking 10.1 hosts,
or at least not guaranteeing not having replication errors (!).

Bug: T274472
Change-Id: I20009f5c38aa1bcc35dff44d64e0764aafce4c97