bigtop: allow to split the HDFS Namenode RPC thread queue
This change adds the support for a dedicated Service RPC queue in the
HDFS Namenode. The idea is to split the default single RPC queue into
two smaller ones: one dedicated to client hadoop RPCs, and the other
dedicated to process HDFS Datanode Block reports and Zookeeper
health checks. It should be used when the number of worker nodes grows
to allow the HDFS Namenode to scale better.

https://community.cloudera.com/t5/Community-Articles/How-do-you-enable-Namenode-service-RPC-port-without-HDFS/ta-p/249006

Bug: T273629
Change-Id: I8ee990f91b0f4152d1ab88c674538206abe9afb2