Nova vendordata: change the mime-type of the cloud-config section
With the current version of cloud-init (20.2), it works fine to
have the mime type set to text/cloud-config.  Future versions
of cloud-init bypass the jinja rendering step for this type;
changing it to text/jinja2 should ensure that this keeps working
with future releases.

Too much info about this can be found in the upstream bug here:

https://bugs.launchpad.net/cloud-init/+bug/1914641

Change-Id: I1eec5ac5e443e1f90d600463fe950c10a098eca1