cloudgw: interfaces: relax check on routing setup by using 'onlink'
As of this writing, we have a /30 for cloud-gw-transport-codfw (vlan 2107) and
therefore we cannot have 1 IP address per device (cloudgw200[12]-dev) plus the
VRRP VIP.

If the device is BACKUP in VRRP/keepalived (so no VIPs configured), when trying
to inject ip routes in the kernel it will complain that the network is
unreachable, with a message similar to this one:

cloudgw2001-dev ifup[1261]: Error: Nexthop has invalid gateway.
cloudgw2001-dev ifup[1261]: ifup: failed to bring up eno2.2107
cloudgw2001-dev ifup[1261]: Error: Nexthop has invalid gateway.
cloudgw2001-dev ifup[1261]: ifup: failed to bring up eno2.2120

Which is kinda true, given we aren't connected to the subnet... just yet

The 'onlink' keyword tells the kernel to don't check if the affected route is
reachable in the current network setup.

Bug: T272963
Change-Id: I411e4fdba9efa8144af54b37be937d2ab84854d4
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>