mediawiki::alerts add alert when 20% of servers is saturated
When LVS is distributing traffing to a mediawiki cluster, it
does not have an overview of which servers are busy and how
busy they are, but it depends on the weight of each server.
This can lead to requests waiting on a busy server, and
increase our overall latency.

There is not an immediate actionable associated with this
alert, but it can be used as an indicator of our servers
getting more busy than usual, which in most cases is related
to some out of the ordinary traffic pattern.

Bug: T267176
Change-Id: I7ffc7535a6f13db2bfdc979fe4694fead3b20575