cloudgw: use address per interface in the cloud-instance-transport subnet
So in the cloud-instance-transport subnet we have a /29 allocated, with enough
empty address to allow using an IP address per interface to run the VRRP
mechanism, instead of running it in the control plane interface.

Note however, that we only have a /30 allocated for the cloud-facing subnet, so
this approach cannot be applied there given we don't have allocation room for
one address per device.

This patch introduces the change for codfw1dev, but eqiad1 also has a /29, so
the config should be same there.

Basically, this is a partial revert of

Bug: T272963
Change-Id: I861d6d169cb89af6c8e0fb5d5e5d5963c9f6dc4b
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>