cloudgw: let keepalived track additional static routes
Given we don't have per-server addresses in the virt-facing subnet, when a
failover happens and the virt-facing VIP is moved to the other server, the
static routes that were previously set in the networking service get lost,
despite the 'onlink' setup.

Let keepalived track these static routes instead, so they are always present in
each server no mather what.

Bug: T272963
Change-Id: I23bdb7f80f0d437139564191c60eee7fc2466f5d
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>