hieradata: introduce 'public_domain' variable
The variable is meant to indicate the "external" or "public" domain under which services are
expected to run. Usually in the form of <service>.<public_domain>. For production the public_domain
is obviously 'wikimedia.org', whereas for staging/testing (e.g. observability) is going to be
some third-level domain, e.g. <foo>.wmcloud.org or <foo>.wmflabs.org.

The idea and intended usage is to be able to stop hardcoding 'wikimedia.org' in configurations and
make realistic testing easier to achieve.

$public_domain is related to, but different than, $domain: the latter refers to the network domain
and can vary between internal and external, whereas the former is a logical/administrative domain
used to host services.

Change-Id: I74ded3d910636a2d89b184110b64400fd116930a