admin: create new group for gitlab-roots
This is creating a new admin group with full root access
to be added on future gitlab VMs which are requested
separately in T274459.

This is only about the group itself and the privileges
it gives and needs SRE meeting approval.

Adding specific people into this group will be done separately
once the names are known and will then just need approval to
add users into an existing group but not the SRE meeting discussion
anymore.

Adding the group to the machines will be done in a placeholder
puppet role which we will have to create even if the VMs are otherwise
not managed by puppet. This will be done after the VMs are created.

Listing Wolfgang as the approver because he is SRE sponsor, could also
be Tyler though.

Bug: T274953
Change-Id: I791185fa484bbc1b07a725575fc7474305874b38