mediawiki: Remove duplicate "in X:Z" from php7-fatal-error.php message
== Throwable vs Unrecoverabe ==

Also update the docs a bit. As the behaviour distinction does not seem
to be caused by postsend vs regular. I'm not sure if this was previously
true or that things have changed since then.

When trying out different ways of producing fatal errors via the
/w/fatal-error.php script, it seems clear that all cases of an uncaught
exception generally don't make it here anymore. They are handled by MW,
and that's it, which is fine. I considered removing handling of this
use case, but then I realized there is actually a way to have
"Uncaught Exception" still go to this script, namely from a destruct
or shutdown callback. It's just that /w/fatal-error.php didn't offer
that feature yet, so I added it in I9a050c4bb0c62953ee77568bd7.

The majority of Logstash reports from php7-fatal-error.php, however,
are from Unrecoverable errors (vs Throwable) and that seems to be the
real indicator of how the trace is transported to this script, so I've
updated the sections to reflect that instead.

== php-wmerrors ==

The previous docs implied that for Unrecoverable errors, the trace is
inserted into the message by php-wmerrors. While php-wmerrors does
have code that appears to be capable of doing it, it uses the separator
"Backtrace:" and not "Stack trace:". I confirmed that for all cases
we know of, this code path in php-wmerrors is inert. This trace is
actually inserted into the message by Zend itself.

== Duplicate info ==

For the rare case of an Uncaught Exception to make it here, the script
now no longer appends a duplicate "in file:line" suffix to the message,
and it also no longer prepends "PHP Fatal Error" to those since they
have "Uncaught Exception" instead. The "PHP Fatal Error" is generally
used as analogous to "PHP Notice" and "PHP Warning" and not for
thrown exceptions that went uncaught. This should make them a bit less
confusing and match local development more closely.

Bug: T275075
Change-Id: Iae0b7561b2219e48246964f123acbbfeacf9f8d2