maintain-meta_p: get case sensitivity from API
As far as I can tell, siprop=general has returned this particular bit of
information since it was first introduced (MediaWiki commit e57335a633),
so I don’t know what that comment was talking about, to be honest.

Note that this is not a no-op change: szywiki has $wgCapitalLinks set to
false in InitialiseSettings.php, but had not been added here. So this
ought to fix the “sensitive” value for that particular wiki.

Change-Id: Ie0d25446aa64c58060e5634e427ebf3365910c33