Update test cluster refine jobs to use event platform schemas
The eventlogging_analytics refine job is to be deprecated.  On the test
cluster, we only refine NavigationTiming, which has been migrated to
event platform schemas, so we should use the eventlogging_legacy job for that.

Also add a refine event job to refine the eventgate-analytics.test.event
stream we are importing.

This is needed to properly test the new Hadoopless Spark package.

Also update to refinery version 0.1.1

Bug: T274384
Change-Id: I60d4e067091f0be1c0573dbe2bc6b5046fd1175f