profile::wmcs::instance: replace hiera_include with lookup
After we removed all hiera() we still have a few special cases
left using other deprecated hiera functions like this one.

I got the "translation" to lookup syntax from:

https://puppet.com/docs/puppet/7.4/hiera_migrate.html#updated_classic_hiera_function_calls

I have to use a lint-ignore because this is a lookup in the middle of
a class. It would have also been downvoted with hiera() but was just
grandfathered in. The style check looks at new additions only.

If you know how to properly fix that too let me know.

Bug: T209953
Change-Id: Idf1e603bc3632826b0dd9374cacc41080dec1a4b